requirements.txt
Flask==2.0.1
Flask-PyMongo==2.3.0
pymongo==3.11.4
Werkzeug==2.0.1
python-dotenv==0.19.2
